refactor: extract checkboxes from maintenance component
