+++ src/pages/maintenance/Checkboxes.module.css
@@ -1,0 +1,7 @@
+.checkboxes {
+    margin-bottom: var(--spacers-dp16);
+}
+
+.checkbox {
+    margin-bottom: var(--spacers-dp4);
+}
