+++ src/pages/maintenance/Maintenance.module.css
@@ -1,11 +1,3 @@
 .card {
     padding: var(--spacers-dp16);
 }
-
-.checkboxes {
-    margin-bottom: var(--spacers-dp16);
-}
-
-.checkbox {
-    margin-bottom: var(--spacers-dp4);
-}
